Suppress some compile warnings

+++ pathfinder/core/scene.cpp
@@ -85,7 +85,7 @@ void Scene::append_scene(const Scene &scene, const Transform2 &transform) {
     auto merged_palette_info = palette.append_palette(scene.palette, transform);
 
     // Merge clip paths.
-    std::vector<size_t> clip_path_mapping;
+    std::vector<uint32_t> clip_path_mapping;
     clip_path_mapping.reserve(scene.clip_paths.size());
     for (auto &clip_path : scene.clip_paths) {
         clip_path_mapping.push_back(clip_paths.size());
@@ -97,7 +97,7 @@ void Scene::append_scene(const Scene &scene, const Transform2 &transform) {
     }
 
     // Merge draw paths.
-    std::vector<size_t> draw_path_mapping;
+    std::vector<uint32_t> draw_path_mapping;
     draw_path_mapping.reserve(scene.draw_paths.size());
     for (auto &draw_path : scene.draw_paths) {
         draw_path_mapping.push_back(draw_paths.size());
